feat: support stream_connection

+++ test/plugin/RawType.h
@@ -3,7 +3,7 @@
 #include "src/serialize.hpp"
 
 struct RawType {
-  int id = 0;
+  uint8_t id = 0;
   std::string name;
   uint8_t age = 0;
 };
